Update instructions to not require second keyboard
By setting a 15 second delay before the `mdloader` is executed, you can flash without having to use a second keyboard. The user will have 15 seconds to hit the reset button on their keyboard, and then the command will run, alleviating the need for running the command with a second keyboard.